Add files via upload
There will be an array of these, one for each 5-minute increment throughout the day. When the user wanted to figure out the parking situation at a particular time the system will called "FindParkingAsOfTime(requestedTime)" the system will then use that time to search through the array of classes for the correct time needed. It will then return one of the classes associated with the time the user queried. This will allow the system to then parse though the class collecting the data about all parking lots and their availability.